+++ src/pages/custom-button.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Text, TouchableOpacity, View } from 'react-native';
 import { general } from '../styles';
 
-export const CustomButton = (props: { title: string; handleClick: () => void }) => {
+export const CustomButton = (props: { title: string; onClick: () => void }) => {
   return (
-    <TouchableOpacity onPress={props.handleClick} style={general.button}>
+    <TouchableOpacity onPress={props.onClick} style={general.button}>
       <View>
         <Text style={{ color: 'white' }}>{props.title}</Text>
       </View>
